Move DataFormat into a top-level class
