+++ x-pack/plugin/inference/src/test/java/org/elasticsearch/xpack/inference/services/jinaai/JinaAIServiceTests.java
@@ -25,6 +25,7 @@ import org.elasticsearch.inference.ChunkInferenceInput;
 import org.elasticsearch.inference.ChunkedInference;
 import org.elasticsearch.inference.ChunkingSettings;
 import org.elasticsearch.inference.ChunkingStrategy;
+import org.elasticsearch.inference.DataFormat;
 import org.elasticsearch.inference.EmbeddingRequest;
 import org.elasticsearch.inference.InferenceService;
 import org.elasticsearch.inference.InferenceServiceConfiguration;
@@ -86,7 +87,7 @@ import java.util.function.Function;
 import static org.elasticsearch.common.xcontent.XContentHelper.toXContent;
 import static org.elasticsearch.inference.DataType.IMAGE;
 import static org.elasticsearch.inference.DataType.TEXT;
-import static org.elasticsearch.inference.InferenceString.DataFormat.BASE64;
+import static org.elasticsearch.inference.DataFormat.BASE64;
 import static org.elasticsearch.inference.ModelConfigurations.SERVICE_SETTINGS;
 import static org.elasticsearch.inference.TaskType.TEXT_EMBEDDING;
 import static org.elasticsearch.test.hamcrest.ElasticsearchAssertions.assertToXContentEquivalent;
@@ -1817,7 +1818,7 @@ public class JinaAIServiceTests extends InferenceServiceTestCase {
             var indexToAdd = randomIntBetween(0, inputs.size() - 1);
             var multipleInferenceStrings = new InferenceStringGroup(
                 List.of(
-                    new InferenceString(TEXT, InferenceString.DataFormat.TEXT, "first_input"),
+                    new InferenceString(TEXT, DataFormat.TEXT, "first_input"),
                     new InferenceString(IMAGE, BASE64, "second_input")
                 )
             );
